# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TimeSlot {
  /** Time in HH:MM format (24h) */
  time: string;
  /** Human-readable label e.g. "9:00 AM" */
  label: string;
  available: boolean;
}

export interface AvailableSlotsResponse {
  date: string;
  slots: TimeSlot[];
}

export interface CreateAppointmentBody {
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  service: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format (24h) */
  timeSlot: string;
  notes?: string;
}

export interface Appointment {
  id: number;
  patientName: string;
  patientEmail: string;
  patientPhone: string;
  service: string;
  date: string;
  timeSlot: string;
  notes?: string;
  status: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  newPassword: string;
}

export interface AdminStats {
  totalBookings: number;
  todayBookings: number;
  pendingBookings: number;
  totalPatients: number;
}

export interface AdminAppointmentsResponse {
  appointments: Appointment[];
  total: number;
  page: number;
  limit: number;
}

export type UpdateAppointmentStatusBodyStatus =
  (typeof UpdateAppointmentStatusBodyStatus)[keyof typeof UpdateAppointmentStatusBodyStatus];

export const UpdateAppointmentStatusBodyStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  pending: "pending",
} as const;

export interface UpdateAppointmentStatusBody {
  status: UpdateAppointmentStatusBodyStatus;
}

export interface RescheduleAppointmentBody {
  /** Date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format (24h) */
  timeSlot: string;
}

export interface PatientSummary {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  appointmentCount: number;
}

export interface AdminPatientsResponse {
  patients: PatientSummary[];
  total: number;
}

export interface PatientProfile {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface AdminPatientDetailResponse {
  patient: PatientProfile;
  appointments: Appointment[];
}

export interface UpdatePatientProfileBody {
  name: string;
}

export interface PatientAppointmentsResponse {
  appointments: Appointment[];
}

export type GetAvailableSlotsParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
  service?: string;
};

export type GetAdminAppointmentsParams = {
  page?: number;
  limit?: number;
  status?: string;
  date?: string;
  service?: string;
  search?: string;
};