# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment booking operations
  - name: auth
    description: Authentication operations
  - name: admin
    description: Admin operations
  - name: patient
    description: Patient self-service operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots
      description: Returns available appointment time slots for a given date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
        - name: service
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlotsResponse"

  /appointments:
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid request or slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Time slot already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new patient account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset link sent (or silently ignored if email not found)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Missing email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password using a valid token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Missing fields or invalid/expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard statistics
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/appointments:
    get:
      operationId: getAdminAppointments
      tags: [admin]
      summary: List all appointments with filtering and pagination
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: service
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAppointmentsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/appointments/{id}:
    patch:
      operationId: updateAdminAppointmentStatus
      tags: [admin]
      summary: Update appointment status
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid ID or status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/appointments/{id}/reschedule:
    patch:
      operationId: rescheduleAdminAppointment
      tags: [admin]
      summary: Reschedule an appointment
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid ID or missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/patients:
    get:
      operationId: getAdminPatients
      tags: [admin]
      summary: List all patients with appointment counts
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPatientsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/patients/{id}:
    get:
      operationId: getAdminPatient
      tags: [admin]
      summary: Get a patient by ID with their appointments
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient with appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPatientDetailResponse"
        "400":
          description: Invalid patient ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patient/me:
    get:
      operationId: getPatientProfile
      tags: [patient]
      summary: Get current patient profile
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePatientProfile
      tags: [patient]
      summary: Update current patient profile name
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientProfile"
        "400":
          description: Missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patient/me/appointments:
    get:
      operationId: getPatientAppointments
      tags: [patient]
      summary: Get all appointments for the current patient
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Patient appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientAppointmentsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /patient/me/appointments/{id}/cancel:
    patch:
      operationId: cancelPatientAppointment
      tags: [patient]
      summary: Cancel an appointment belonging to the current patient
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid ID or appointment cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AvailableSlotsResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - date
        - slots

    TimeSlot:
      type: object
      properties:
        time:
          type: string
          description: Time in HH:MM format (24h)
        label:
          type: string
          description: Human-readable label e.g. "9:00 AM"
        available:
          type: boolean
      required:
        - time
        - label
        - available

    CreateAppointmentBody:
      type: object
      properties:
        patientName:
          type: string
        patientEmail:
          type: string
        patientPhone:
          type: string
        service:
          type: string
        date:
          type: string
          description: Date in YYYY-MM-DD format
        timeSlot:
          type: string
          description: Time in HH:MM format (24h)
        notes:
          type: string
      required:
        - patientName
        - patientEmail
        - patientPhone
        - service
        - date
        - timeSlot

    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientName:
          type: string
        patientEmail:
          type: string
        patientPhone:
          type: string
        service:
          type: string
        date:
          type: string
        timeSlot:
          type: string
        notes:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - patientName
        - patientEmail
        - patientPhone
        - service
        - date
        - timeSlot
        - status
        - createdAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - name
        - email
        - role

    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - name
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required:
        - token
        - newPassword

    AdminStats:
      type: object
      properties:
        totalBookings:
          type: integer
        todayBookings:
          type: integer
        pendingBookings:
          type: integer
        totalPatients:
          type: integer
      required:
        - totalBookings
        - todayBookings
        - pendingBookings
        - totalPatients

    AdminAppointmentsResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - appointments
        - total
        - page
        - limit

    UpdateAppointmentStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled, completed, pending]
      required:
        - status

    RescheduleAppointmentBody:
      type: object
      properties:
        date:
          type: string
          description: Date in YYYY-MM-DD format
        timeSlot:
          type: string
          description: Time in HH:MM format (24h)
      required:
        - date
        - timeSlot

    PatientSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
        appointmentCount:
          type: integer
      required:
        - id
        - name
        - email
        - createdAt
        - appointmentCount

    AdminPatientsResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/PatientSummary"
        total:
          type: integer
      required:
        - patients
        - total

    PatientProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt

    AdminPatientDetailResponse:
      type: object
      properties:
        patient:
          $ref: "#/components/schemas/PatientProfile"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required:
        - patient
        - appointments

    UpdatePatientProfileBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    PatientAppointmentsResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required:
        - appointments