# lib/api-zod/src/generated/types/adminAppointmentsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";

export interface AdminAppointmentsResponse {
  appointments: Appointment[];
  total: number;
  page: number;
  limit: number;
}