# lib/api-zod/src/generated/types/adminPatientDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";
import type { PatientProfile } from "./patientProfile";

export interface AdminPatientDetailResponse {
  patient: PatientProfile;
  appointments: Appointment[];
}