# lib/api-zod/src/generated/types/patientProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PatientProfile {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}